Explore softsubs to see if they're feasible
When concatenating videos A and B, len(cat(A, B)) != len(A) + len(B).
There's usually a difference of at least a few hundredths of a second,
at least according to format.duration from ffprobe. I don't see any way
to automatically find the right timings for long compilations. Also, I
doubt many, if any, media players support subtitles with emoji out of
the box. For VLC to display them I needed to create a font that
contained both emoji and text characters. Maybe using a more advanced
subtitle format and embedding the right fonts would work, but I doubt
that would be useful for any web-based video hosting services.

Whereas the only problem hardsubs have is that libass doesn't seem to use
fontconfig properly and the font can't be chosen. But it grabs
characters it needs wherever it can find them, so as long as an emoji
font is installed the characters will show up, at least.