Fix moving scaled files
Remove the special case so TestScale covers moveFile better.

It was awkward to test moving files from one mount to another, so I
didn't. The argument order was wrong for both moveFile and os.Rename,
which cancelled out for os.Rename, but not for copying by bytes, which
was used for cross-mount copies. This resulted in vines to be scaled
being deleted instead.